backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
faiss-cpu==1.7.4
google-generativeai==0.3.2
numpy==1.24.3
gunicorn==21.2.0
python-dotenv
fastapi
uvicorn[standard]
pydantic